+++ src/components/todo/AuthendicateRoute.jsx
@@ -1,0 +1,8 @@
+import {Component} from 'react';
+import Authenticate from "./Authenticate";
+
+class AuthendicateRoute  extends Component{
+    render() {
+
+    }
+}
