+++ src/components/todo/FooderComponet.jsx
@@ -1,0 +1,13 @@
+import React, {Component} from 'react';
+
+class FooderComponet extends Component {
+    render() {
+        return(
+            <footer className="footer">
+                All content Reserved for Ramesh.
+            </footer>
+        )
+    }
+}
+
+export default FooderComponet;
